+++ tests/ethernet/test_ethernet_network.py
@@ -1,7 +1,7 @@
 import os
 import socket
 import time
-from ftplib import error_perm, error_temp
+from ftplib import error_temp
 from threading import Thread
 
 import pytest
@@ -27,38 +27,50 @@ class FTPServer(Thread):
     """
 
     def __init__(
-        self, folder_path: str = "./", new_user: str = "user", new_password: str = "password"
+        self,
+        folder_path: str = "./",
+        new_user: str = "user",
+        new_password: str = "password",
     ):
         super().__init__()
+        self.ftp_port = 21
         self.fpt_checker = InMemoryUsernamePasswordDatabaseDontUse()
         self.fpt_checker.addUser(new_user, new_password)
         self.ftp_portal = Portal(
             FTPRealm(folder_path, folder_path), [AllowAnonymousAccess(), self.fpt_checker]
         )
         self.ftp_factory = FTPFactory(self.ftp_portal)
-        reactor.listenTCP(21, self.ftp_factory)
+        self.reactor = reactor
+        self.__stopped = False
+        self.reactor.listenTCP(self.ftp_port, self.ftp_factory)
 
     def run(self) -> None:
         """Run FTP server."""
-        reactor.run()
+        self.reactor.run(installSignalHandlers=False)
 
     def stop(self) -> None:
         """Stop FTP server."""
-        reactor.stop()
+        if self.__stopped:
+            return
+        self.__stopped = True
+        self.reactor.callFromThread(self.reactor.stop)
+
+    def join(self, timeout=None):
+        self.stop()
+        return super().join(timeout)
 
 
-@pytest.fixture()
-def ftp_server_manager(request):
+@pytest.fixture(scope="module")
+def ftp_server_manager():
     # Get configuration
-    folder_path = request.param.get("folder_path", "./")
-    ftp_user = request.param.get("ftp_user", "user")
-    ftp_password = request.param.get("ftp_password", "password")
+    ftp_user = "user"
+    ftp_password = "password"
     # Create FTP server
-    server = FTPServer(folder_path=folder_path, new_user=ftp_user, new_password=ftp_password)
+    server = FTPServer(folder_path="./", new_user=ftp_user, new_password=ftp_password)
     server.start()
-    yield folder_path, ftp_user, ftp_password
-    server.stop()
+    yield ftp_user, ftp_password
     server.join()
+    assert not server.is_alive()
 
 
 @pytest.fixture()
@@ -166,43 +178,36 @@ def test_load_firmware_no_connection():
     os.remove(fw_file)
 
 
-@pytest.mark.skip(reason="Skipping for now. Check INGK-1035 & INGK-1058.")
 @pytest.mark.no_connection
-@pytest.mark.parametrize(
-    "ftp_server_manager",
-    [{"folder_path": "./", "ftp_user": "user", "ftp_password": "password"}],
-    indirect=True,
-)
 def test_load_firmware_wrong_user_pwd(ftp_server_manager):
     """Testing failed ftp firmware load with fake FTP server."""
     fw_file = "temp_file.lfu"
     with open(fw_file, "w"):
         pass
-    folder_path, ftp_user, ftp_password = ftp_server_manager
+    ftp_user, ftp_password = ftp_server_manager
     # Wrong user and password
     fake_user = "mamma"
     fake_password = "mia"
     # Create Network
     net = EthernetNetwork()
     with pytest.raises(ILFirmwareLoadError) as excinfo:
-        net.load_firmware(fw_file, target="localhost", ftp_user=fake_user, ftp_pwd=fake_password)
+        net.load_firmware(
+            fw_file,
+            target="localhost",
+            ftp_user=fake_user,
+            ftp_pwd=fake_password,
+        )
     assert str(excinfo.value) == "Unable to login the FTP session"
-    assert isinstance(excinfo.value.__cause__, error_perm)
     os.remove(fw_file)
 
 
 @pytest.mark.no_connection
-@pytest.mark.parametrize(
-    "ftp_server_manager",
-    [{"folder_path": "./", "ftp_user": "user", "ftp_password": "password"}],
-    indirect=True,
-)
 def test_load_firmware_error_during_loading(mocker, ftp_server_manager):
     """Testing failed ftp firmware load with fake FTP server."""
     fw_file = "temp_file.lfu"
     with open(fw_file, "w"):
         pass
-    folder_path, ftp_user, ftp_password = ftp_server_manager
+    ftp_user, ftp_password = ftp_server_manager
     net = EthernetNetwork()
     # Mock ftp error for ftp.stobinary call
     mocker.patch(